refactor: move home_page.dart to feature/home/ui
