+++ apps/app/lib/main.dart
@@ -1,10 +1,7 @@
-import 'dart:async';
-
 import 'package:flutter/material.dart';
-import 'package:flutter_app/data/provider/api/repository_api.dart';
 import 'package:flutter_app/data/provider/shared_preferences.dart';
+import 'package:flutter_app/feature/home/ui/home_page.dart';
 import 'package:flutter_app/feature/setting/provider/theme_mode_notifier.dart';
-import 'package:flutter_app/feature/setting/setting_page.dart';
 import 'package:flutter_app/util/logger.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:package_info_plus/package_info_plus.dart';
@@ -16,12 +13,12 @@ void main() async {
   final packageInfo = await PackageInfo.fromPlatform();
   logger.info(packageInfo);
 
-  final sharedPreferences = await SharedPreferences.getInstance();
+  final prefs = await SharedPreferences.getInstance();
 
   runApp(
     ProviderScope(
       overrides: [
-        sharedPreferencesProvider.overrideWithValue(sharedPreferences),
+        sharedPreferencesProvider.overrideWithValue(prefs),
       ],
       child: const MainApp(),
     ),
@@ -40,51 +37,6 @@ class MainApp extends ConsumerWidget {
       darkTheme: ThemeData.dark(),
       themeMode: themeMode,
       home: const HomePage(),
-    );
-  }
-}
-
-class HomePage extends ConsumerWidget {
-  const HomePage({super.key});
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final repositories = ref.watch(listPublicRepositoriesProvider);
-
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('ホーム'),
-        actions: [
-          IconButton(
-            onPressed: () {
-              unawaited(
-                Navigator.of(context).push(
-                  MaterialPageRoute<void>(
-                    builder: (context) {
-                      return const SettingPage();
-                    },
-                  ),
-                ),
-              );
-            },
-            icon: const Icon(
-              Icons.settings,
-            ),
-          ),
-        ],
-      ),
-      body: Center(
-        child: repositories.when(
-          data: (value) => ListView.builder(
-            itemCount: value.length,
-            itemBuilder: (context, index) => ListTile(
-              title: Text(value[index].name),
-            ),
-          ),
-          error: (error, _) => Text(error.toString()),
-          loading: CircularProgressIndicator.new,
-        ),
-      ),
     );
   }
 }
